% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propfit.R
\name{propfit}
\alias{propfit}
\title{Fit Proportions.}
\usage{
propfit(
  y,
  x = NULL,
  start = NULL,
  weights = NULL,
  offset = NULL,
  ...,
  estfun = FALSE,
  object = FALSE
)
}
\arguments{
\item{y}{A matrix or data.frame where each row corresponds to a k-dim observation.}

\item{x}{Not used.}

\item{start}{Not used.}

\item{weights}{Not used yet.}

\item{offset}{Not used.}

\item{...}{Not used yet.}

\item{estfun}{logical. Should the matrix of score contributions (estimating
functions) be returned.}

\item{object}{logical. Should the return from `optim()` be returned.}
}
\description{
Maximum likelihood estimation of Dirichlet parameter.
}
\details{
Perform MLE of parameters of a Dirichlet distribution.
Interface and return value are designed to fit
the modelparty infrastructure of the partykit package.
ML estimates are found numerically using `optim` with
BFGS algorithm.
}
