% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proptree.R
\name{proptree}
\alias{proptree}
\alias{print.proptree}
\alias{plot.proptree}
\title{Proportional Trees.}
\usage{
proptree(formula, data, na.action, ...)

\method{print}{proptree}(x, title = "Proportion tree", objfun = "negative log-likelihood", ...)

\method{plot}{proptree}(x, proportions = TRUE, ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit. This
should either be of type `y1 + y2 + y3 ~ x1 + x2` with reponse
vectors `y1`, `y2`, and `y3` or `y ~ x1 + x2`
with a matrix response `y`. `x1` and `x2` are used as
partitioning variables.}

\item{data}{A data frame containing the variables in the model.}

\item{na.action}{A function which indicates what should happen when the data
contain missing values `NA`.}

\item{...}{arguments passed to [partykit::mob_control()],
[partykit::print.modelparty()] and [partykit::plot.modelparty()].}

\item{x}{An object of class `proptree`.}

\item{title}{character, title for print method.}

\item{objfun}{character, labeling objective function.}

\item{proportions}{logical, display proportions (default) or
log-alphas of Dirichlet distribution.}
}
\value{
An object of class `proptree` inheriting from `modelparty`.
}
\description{
Fit tree models based on the Dirichlet distribution.
}
\details{
Fit tree models on proportional or compositional data
based on the Dirichlet distribution. This funciton
is a high-level interface to `partykit::mob()` in combination
with the fitter `propfit()`.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: print method

\item \code{plot}: plot method
}}

\examples{
   ### --- toy data ---
   set.seed(111)
   n <- 500
   x <- sample(c(-1, 1), n, replace = TRUE)
   alpha <- cbind(rnorm(n, 1.4, sd = .1), rnorm(n, 1.4, sd = .1), jitter(2 + x))
   y <- alpha / rowSums(alpha)
   d <- as.data.frame(y)
   names(d) <- paste0("y", 1:3)
   d$x <- jitter(x)
   tr <- proptree(y1 + y2 + y3 ~ x, data = d)
   plot(tr)

   ### --- Sediment composition  ---
   data("ArcticLake", package = "DirichletReg")
   tr <- proptree(sand + silt + clay ~ depth, data = ArcticLake)
   plot(tr)

}
\seealso{
[partykit::mob()], [propfit()]
}
